Struct zero_sum::analysis::search::Analysis

pub struct Analysis<'a, E, S, P, R> where E: Evaluation,
        S: 'a + State<P, R> + Evaluatable<E> + Extrapolatable<P>,
        P: Ply,
        R: Resolution {
    pub state: &'a S,
    pub evaluation: E,
    pub principal_variation: Vec<P>,
    pub stats: Option<Box<Display>>,
    // some fields omitted
}

The result of a search.

Fields
state: &'a S

A reference to the state on which the search was performed.

evaluation: E

The evaluation of the state after applying the principal variation.

principal_variation: Vec<P>

The principal variation of the state.

stats: Option<Box<Display>>

Optional statistics from the search may be available for printing.