Struct zero_sum::analysis::search::PvSearch

pub struct PvSearch<E, S, P, R> where E: Evaluation, S: State<P, R> + Evaluatable<E> + Extrapolatable<P>, P: Ply, R: Resolution {
    // some fields omitted
}

A PVS implementation of Search with a few common optimizations.

Example

let state = State::new();
let (interrupt_sender, interrupt_receiver) = std::sync::mpsc::channel();

let mut search = PvSearch::<Eval, State, Ply, Resolution>::with_depth(5);
let analysis = search.search(&state, Some(interrupt_receiver));

Methods

impl<E, S, P, R> PvSearch<E, S, P, R> where E: Evaluation, S: State<P, R> + Evaluatable<E> + Extrapolatable<P>, P: Ply, R: Resolution

fn new() -> PvSearch<E, S, P, R>

Creates a PvSearch without a target depth or time goal. It will search until it finds a favorable resolution, or until the search is interrupted.

fn with_depth(depth: u8) -> PvSearch<E, S, P, R>

Creates a PvSearch that will search to a maximum depth of depth.

fn with_goal(goal: u16, branching_factor: f32) -> PvSearch<E, S, P, R>

Creates a PvSearch that will search until it predicts that it will exceed goal seconds with the next depth. branching_factor is used to predict the required time to search at the next depth.